set 0.2 to float instead of casting

+++ src/main/java/org/violetmoon/quark/content/mobs/client/model/ToretoiseModel.java
@@ -282,7 +282,7 @@ public class ToretoiseModel extends EntityModel<Toretoise> {
 
 			trans *= (1F - finalRideMultiplier);
 			rot *= (1F - finalRideMultiplier);
-			trans += (float) (finalRideMultiplier * -0.2);
+			trans += finalRideMultiplier * -0.2F;
 
 			matrix.pushPose();
